Switch out curl_fetch_web_data for Guzzle in the avatar-sizing routine and solidify them all into a single global method.
